refactor: dedupe game/scene event handling
