+++ internal/game/game.go
@@ -10,8 +10,6 @@ package game
 import (
 	"fmt"
 	"image/color"
-	"path/filepath"
-	"time"
 
 	"github.com/hajimehoshi/ebiten/v2"
 	"github.com/hajimehoshi/ebiten/v2/ebitenutil"
@@ -174,149 +172,7 @@ func (g *Game) Layout(outsideWidth, outsideHeight int) (int, int) {
 
 // handleClaudeEvent processes events from the Claude interceptor
 func (g *Game) handleClaudeEvent(event *claude.Event) {
-	switch event.Type {
-	case claude.EventFileRead:
-		// Reveal fog for files that Claude reads
-		if path, ok := event.Data["file_path"].(string); ok {
-			absPath := filepath.Join(g.projectPath, path)
-			g.mapView.RevealTile(absPath)
-		}
-
-	case claude.EventFileWrite, claude.EventFileEdit:
-		// Highlight files that Claude writes/edits
-		if path, ok := event.Data["file_path"].(string); ok {
-			absPath := filepath.Join(g.projectPath, path)
-			// Reveal and highlight the tile
-			g.mapView.RevealTile(absPath)
-			// Trigger advisors that watch this file (Phase 3)
-			g.triggerAdvisorsForFile(path)
-		}
-
-	case claude.EventBuildRun:
-		// Update build status in resource tracker
-		// TODO: Extract build results and update resources
-
-	case claude.EventTestRun:
-		// Update test results
-		// TODO: Extract test results and update resources
-
-	case claude.EventSubagentRun:
-		// Handle advisor/subagent execution (Phase 3)
-		g.handleAdvisorEvent(event)
-	}
-}
-
-// handleAdvisorEvent processes advisor-related events
-func (g *Game) handleAdvisorEvent(event *claude.Event) {
-	// Extract advisor ID from event data if present
-	advisorID, _ := event.Data["advisor_id"].(string)
-	if advisorID == "" {
-		return
-	}
-
-	adv := g.advisorPool.Get(advisorID)
-	if adv == nil {
-		return
-	}
-
-	// Check if this is a start or completion event
-	if status, ok := event.Data["status"].(string); ok {
-		switch status {
-		case "started":
-			adv.StartAnalysis()
-		case "completed":
-			duration, _ := event.Data["duration_ms"].(float64)
-			tokensIn, _ := event.Data["tokens_in"].(float64)
-			tokensOut, _ := event.Data["tokens_out"].(float64)
-			adv.CompleteAnalysis(
-				durationFromMs(duration),
-				int64(tokensIn),
-				int64(tokensOut),
-				nil,
-			)
-		case "error":
-			errMsg, _ := event.Data["error"].(string)
-			adv.CompleteAnalysis(0, 0, 0, fmt.Errorf("%s", errMsg))
-		}
-	}
-
-	// Check for insights in the event
-	if insightsData, ok := event.Data["insights"].([]interface{}); ok {
-		for _, insightData := range insightsData {
-			if insightMap, ok := insightData.(map[string]interface{}); ok {
-				insight := g.parseInsight(advisorID, insightMap)
-				if insight != nil {
-					adv.AddInsight(insight)
-				}
-			}
-		}
-	}
-}
-
-// parseInsight parses insight data from an event
-func (g *Game) parseInsight(advisorID string, data map[string]interface{}) *advisor.Insight {
-	title, _ := data["title"].(string)
-	description, _ := data["description"].(string)
-	if title == "" {
-		return nil
-	}
-
-	// Map severity
-	severityStr, _ := data["severity"].(string)
-	severity := advisor.SeverityInfo
-	switch severityStr {
-	case "warning":
-		severity = advisor.SeverityWarning
-	case "critical":
-		severity = advisor.SeverityCritical
-	}
-
-	// Map category
-	categoryStr, _ := data["category"].(string)
-	category := advisor.CategoryGeneral
-	switch categoryStr {
-	case "security":
-		category = advisor.CategorySecurity
-	case "performance":
-		category = advisor.CategoryPerformance
-	case "refactoring":
-		category = advisor.CategoryRefactoring
-	case "testing":
-		category = advisor.CategoryTesting
-	}
-
-	insight := advisor.NewInsight(advisorID, title, description, severity, category)
-
-	// Add location if present
-	if filePath, ok := data["file_path"].(string); ok {
-		line, _ := data["line"].(float64)
-		column, _ := data["column"].(float64)
-		insight.WithLocation(filePath, int(line), int(column))
-	}
-
-	// Add suggestion if present
-	if suggestion, ok := data["suggestion"].(string); ok {
-		codeBefore, _ := data["code_before"].(string)
-		codeAfter, _ := data["code_after"].(string)
-		insight.WithSuggestion(suggestion, codeBefore, codeAfter)
-	}
-
-	return insight
-}
-
-// triggerAdvisorsForFile triggers advisors that should run when a file changes
-func (g *Game) triggerAdvisorsForFile(filePath string) {
-	triggered := g.advisorPool.TriggerOnFileChange(filePath)
-	for _, adv := range triggered {
-		// In a real implementation, this would spawn the advisor subagent
-		// For now, we just mark that the advisor should analyze this file
-		_ = adv // TODO: Implement actual advisor execution
-	}
-}
-
-// durationFromMs converts milliseconds to time.Duration
-func durationFromMs(ms float64) time.Duration {
-	return time.Duration(ms * float64(time.Millisecond))
+	processClaudeEvent(event, g.projectPath, g.mapView, g.advisorPool)
 }
 
 // Close cleans up game resources
